Jkr laptop

+++ app.js
@@ -4,6 +4,7 @@ const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose")
 const date = require(__dirname + "/date.js");
+const _ = require("lodash")
 
 const app = express();
 
@@ -44,13 +45,14 @@ const listSchema = new mongoose.Schema({
 // Create the list model
 const List = mongoose.model("List", listSchema)
 
+// Load root - default documents
 app.get("/", function(req, res) {
   const day = date.getDate();
+  // Get all documents
   Item.find()
     .then(result => {
       if(result.length === 0){
-        // Add the documents to the collection
-        // (save the records to the table)
+        // If not documents, add default document and load the page again
         Item.insertMany(defaultItems)
           .then(() => {
             console.log("Default documents have been added to the collection")
@@ -60,21 +62,26 @@ app.get("/", function(req, res) {
           })
           res.redirect("/")
       } else {
-        res.render("list", {listTitle: day, newListItems: result})
+        // Render list.ejs with all documents
+        res.render("list", {listTitle: "Today", newListItems: result})
       }
     })
     .catch(err => {console.log(err)})
 });
 
+// Get custom documents based on url parameter
 app.get("/:customListName", (req, res) => {
   const day = date.getDate();
-  const customListName = req.params.customListName
+  const customListName = _.capitalize(req.params.customListName)
   
+  // Check DB and see if there is a collection for customListName
   List.findOne({name: customListName})
     .then((result) => {
       if(result){
+        // If there is a collection, render list.ejs with collectionList documents
         res.render("list", {listTitle: result.name, newListItems: result.items})
       } else {
+        // No collection, create one and add a default task. Refresh the page
         const list = new List({
           name: customListName,
           items: defaultItems
@@ -88,57 +95,53 @@ app.get("/:customListName", (req, res) => {
     })
 })
 
+// Post a new task to a collection
 app.post("/", function(req, res){
-
   const listName = req.body.list
-  console.log(listName)
 
-  
-
-  List.findOne({name: listName})
-    .then((result) => {
-      if(result){
+  if(listName === "Today"){
+    // If collection = Today, add item to default list
+    new Item({name: req.body.newItem}).save()
+    res.redirect("/")
+  } else {
+    // Find the collection for listName
+    List.findOne({name: listName})
+      .then((result) => {
+        // If collection exists, add new item to the result and load the /<collectionName> url
         result.items.push(new Item({name: req.body.newItem}))
         result.save()
         res.redirect("/"+listName)
-      } else {
-        new Item({name: req.body.newItem}).save()
-        res.redirect("/")
-      }
-    })
+      })
+      .catch(err => {
+        console.log(err)
+      })
+  }
 
-  // new Item({
-  //   name: req.body.newItem
-  // }).save()
-
-  // if (req.body.list === "Work") {
-  //   workItems.push(item);
-  //   res.redirect("/work");
-  // } else {
-  //   items.push(item);
-  //   res.redirect("/");
-  // }
 });
 
+// Delete item from collection
 app.post("/delete", function(req, res){
+  // Get item id from checkbox
   const checkedItem = req.body.checkBox
-  console.log(checkedItem)
-  Item.findByIdAndRemove(checkedItem)
-  .then(result =>{
-    res.redirect("/")
-  })
-  .catch(err => {
-    console.log(err)
-  })
+  // Get collection name from hidden input
+  const hiddenListVal = req.body.hiddenListName
+
+  if(hiddenListVal === "Today"){
+    // Delete from default items list
+    Item.findByIdAndRemove(checkedItem)
+      .then(result =>{res.redirect("/")})
+      .catch(err => {console.log(err)})
+  } else {
+    // Find the collections name
+    List.findOneAndUpdate({name: hiddenListVal},
+      {$pull: {items: {_id: checkedItem}}})
+      .then(result => {
+        console.log("Document removed from collection")
+        res.redirect("/"+hiddenListVal)
+      })
+      .catch(err => {console.log(err)})
+  }
 })
-
-app.get("/work", function(req,res){
-  res.render("list", {listTitle: "Work List", newListItems: workItems});
-});
-
-app.get("/about", function(req, res){
-  res.render("about");
-});
 
 app.listen(3000, function() {
   console.log("Server started on port 3000");
